Add files via upload
This is a windows based game where a player controls a dot and square on a bordered plane. As it moves forward, it leaves a trail behind, resembling a moving snake. In this game the end ofthe trail is in a fixed position, so the snake continually gets longer as it moves.